sensors/oxigeno/requirements.txt
flask
requests
paho-mqtt
# y cualquier otra dependencia que necesites
